move searx.shared.redisdb to searx.redisdb
