--- searx/shared/__init__.py
@@ -1,6 +1,0 @@
-# SPDX-License-Identifier: AGPL-3.0-or-later
-# lint: pylint
-"""Initialization of a *shared* storage.
-"""
-
-from . import redisdb
